Improve GetStepTowardsBSP

+++ blakserv/roofile.c
@@ -986,9 +986,10 @@ bool BSPGetStepTowards(room_type* Room, V2* S, V2* E, V2* P, unsigned int* Flags
    // note: we must verify the location the object is actually going to end up in KOD,
    // this means we must round to the next closer kod-fineness value,  
    // so these values are also exactly expressable in kod coordinates.
+   // in fact this makes the vector a variable length between ~15.5 and ~16.5 fine units
    V2ADD(&stepend, S, &se);
-   stepend.X = (float)ROUNDROOTOKODFINENESS(stepend.X);
-   stepend.Y = (float)ROUNDROOTOKODFINENESS(stepend.Y);
+   stepend.X = ROUNDROOTOKODFINENESS(stepend.X);
+   stepend.Y = ROUNDROOTOKODFINENESS(stepend.Y);
    if (BSPCanMoveInRoom(Room, S, &stepend))
    {
       *P = stepend;
@@ -996,7 +997,7 @@ bool BSPGetStepTowards(room_type* Room, V2* S, V2* E, V2* P, unsigned int* Flags
       *Flags &= ~ESTATE_CLOCKWISE;
       return true;
    }
-
+   
    /****************************************************/
    // 2) can't do direct step
    /****************************************************/
@@ -1020,8 +1021,8 @@ bool BSPGetStepTowards(room_type* Room, V2* S, V2* E, V2* P, unsigned int* Flags
 		 // try 22.5° right
 		 V2ROTATE(&v, 0.5f * (float)-M_PI_4);
 		 V2ADD(&stepend, S, &v);
-		 stepend.X = (float)ROUNDROOTOKODFINENESS(stepend.X);
-		 stepend.Y = (float)ROUNDROOTOKODFINENESS(stepend.Y);
+         stepend.X = ROUNDROOTOKODFINENESS(stepend.X);
+         stepend.Y = ROUNDROOTOKODFINENESS(stepend.Y);
 		 if (BSPCanMoveInRoom(Room, S, &stepend))
 		 {
             *P = stepend;
@@ -1033,8 +1034,8 @@ bool BSPGetStepTowards(room_type* Room, V2* S, V2* E, V2* P, unsigned int* Flags
          // try 45° right
          V2ROTATE(&v, 0.5f * (float)-M_PI_4);
          V2ADD(&stepend, S, &v);
-		 stepend.X = (float)ROUNDROOTOKODFINENESS(stepend.X);
-		 stepend.Y = (float)ROUNDROOTOKODFINENESS(stepend.Y);
+         stepend.X = ROUNDROOTOKODFINENESS(stepend.X);
+         stepend.Y = ROUNDROOTOKODFINENESS(stepend.Y);
          if (BSPCanMoveInRoom(Room, S, &stepend))
          {
             *P = stepend;
@@ -1046,8 +1047,8 @@ bool BSPGetStepTowards(room_type* Room, V2* S, V2* E, V2* P, unsigned int* Flags
          // try 67.5° right
          V2ROTATE(&v, 0.5f * (float)-M_PI_4);
          V2ADD(&stepend, S, &v);
-         stepend.X = (float)ROUNDROOTOKODFINENESS(stepend.X);
-         stepend.Y = (float)ROUNDROOTOKODFINENESS(stepend.Y);
+         stepend.X = ROUNDROOTOKODFINENESS(stepend.X);
+         stepend.Y = ROUNDROOTOKODFINENESS(stepend.Y);
          if (BSPCanMoveInRoom(Room, S, &stepend))
          {
             *P = stepend;
@@ -1059,8 +1060,8 @@ bool BSPGetStepTowards(room_type* Room, V2* S, V2* E, V2* P, unsigned int* Flags
          // try 90° right
          V2ROTATE(&v, 0.5f * (float)-M_PI_4);
          V2ADD(&stepend, S, &v);
-		 stepend.X = (float)ROUNDROOTOKODFINENESS(stepend.X);
-		 stepend.Y = (float)ROUNDROOTOKODFINENESS(stepend.Y);
+         stepend.X = ROUNDROOTOKODFINENESS(stepend.X);
+         stepend.Y = ROUNDROOTOKODFINENESS(stepend.Y);
          if (BSPCanMoveInRoom(Room, S, &stepend))
          {
             *P = stepend;
@@ -1072,8 +1073,8 @@ bool BSPGetStepTowards(room_type* Room, V2* S, V2* E, V2* P, unsigned int* Flags
          // try 112.5° right
          V2ROTATE(&v, 0.5f * (float)-M_PI_4);
          V2ADD(&stepend, S, &v);
-         stepend.X = (float)ROUNDROOTOKODFINENESS(stepend.X);
-         stepend.Y = (float)ROUNDROOTOKODFINENESS(stepend.Y);
+         stepend.X = ROUNDROOTOKODFINENESS(stepend.X);
+         stepend.Y = ROUNDROOTOKODFINENESS(stepend.Y);
          if (BSPCanMoveInRoom(Room, S, &stepend))
          {
             *P = stepend;
@@ -1085,8 +1086,8 @@ bool BSPGetStepTowards(room_type* Room, V2* S, V2* E, V2* P, unsigned int* Flags
          // try 135° right
          V2ROTATE(&v, (float)-M_PI_4);
          V2ADD(&stepend, S, &v);
-		 stepend.X = (float)ROUNDROOTOKODFINENESS(stepend.X);
-		 stepend.Y = (float)ROUNDROOTOKODFINENESS(stepend.Y);
+         stepend.X = ROUNDROOTOKODFINENESS(stepend.X);
+         stepend.Y = ROUNDROOTOKODFINENESS(stepend.Y);
          if (BSPCanMoveInRoom(Room, S, &stepend))
          {
             *P = stepend;
@@ -1107,8 +1108,8 @@ bool BSPGetStepTowards(room_type* Room, V2* S, V2* E, V2* P, unsigned int* Flags
          // try 22.5° left
          V2ROTATE(&v, 0.5f * (float)M_PI_4);
          V2ADD(&stepend, S, &v);
-         stepend.X = (float)ROUNDROOTOKODFINENESS(stepend.X);
-         stepend.Y = (float)ROUNDROOTOKODFINENESS(stepend.Y);
+         stepend.X = ROUNDROOTOKODFINENESS(stepend.X);
+         stepend.Y = ROUNDROOTOKODFINENESS(stepend.Y);
          if (BSPCanMoveInRoom(Room, S, &stepend))
          {
             *P = stepend;
@@ -1120,8 +1121,8 @@ bool BSPGetStepTowards(room_type* Room, V2* S, V2* E, V2* P, unsigned int* Flags
          // try 45° left
          V2ROTATE(&v, 0.5f * (float)M_PI_4);
          V2ADD(&stepend, S, &v);
-		 stepend.X = (float)ROUNDROOTOKODFINENESS(stepend.X);
-		 stepend.Y = (float)ROUNDROOTOKODFINENESS(stepend.Y);
+         stepend.X = ROUNDROOTOKODFINENESS(stepend.X);
+         stepend.Y = ROUNDROOTOKODFINENESS(stepend.Y);
          if (BSPCanMoveInRoom(Room, S, &stepend))
          {
             *P = stepend;
@@ -1133,8 +1134,8 @@ bool BSPGetStepTowards(room_type* Room, V2* S, V2* E, V2* P, unsigned int* Flags
          // try 67.5° left
          V2ROTATE(&v, 0.5f * (float)M_PI_4);
          V2ADD(&stepend, S, &v);
-         stepend.X = (float)ROUNDROOTOKODFINENESS(stepend.X);
-         stepend.Y = (float)ROUNDROOTOKODFINENESS(stepend.Y);
+         stepend.X = ROUNDROOTOKODFINENESS(stepend.X);
+         stepend.Y = ROUNDROOTOKODFINENESS(stepend.Y);
          if (BSPCanMoveInRoom(Room, S, &stepend))
          {
             *P = stepend;
@@ -1146,8 +1147,8 @@ bool BSPGetStepTowards(room_type* Room, V2* S, V2* E, V2* P, unsigned int* Flags
          // try 90° left
          V2ROTATE(&v, 0.5f * (float)M_PI_4);
          V2ADD(&stepend, S, &v);
-		 stepend.X = (float)ROUNDROOTOKODFINENESS(stepend.X);
-		 stepend.Y = (float)ROUNDROOTOKODFINENESS(stepend.Y);
+         stepend.X = ROUNDROOTOKODFINENESS(stepend.X);
+         stepend.Y = ROUNDROOTOKODFINENESS(stepend.Y);
          if (BSPCanMoveInRoom(Room, S, &stepend))
          {
             *P = stepend;
@@ -1159,8 +1160,8 @@ bool BSPGetStepTowards(room_type* Room, V2* S, V2* E, V2* P, unsigned int* Flags
          // try 112.5° left
          V2ROTATE(&v, 0.5f * (float)M_PI_4);
          V2ADD(&stepend, S, &v);
-         stepend.X = (float)ROUNDROOTOKODFINENESS(stepend.X);
-         stepend.Y = (float)ROUNDROOTOKODFINENESS(stepend.Y);
+         stepend.X = ROUNDROOTOKODFINENESS(stepend.X);
+         stepend.Y = ROUNDROOTOKODFINENESS(stepend.Y);
          if (BSPCanMoveInRoom(Room, S, &stepend))
          {
             *P = stepend;
@@ -1172,8 +1173,8 @@ bool BSPGetStepTowards(room_type* Room, V2* S, V2* E, V2* P, unsigned int* Flags
          // try 135° left
          V2ROTATE(&v, 0.5f * (float)M_PI_4);
          V2ADD(&stepend, S, &v);
-		 stepend.X = (float)ROUNDROOTOKODFINENESS(stepend.X);
-		 stepend.Y = (float)ROUNDROOTOKODFINENESS(stepend.Y);
+         stepend.X = ROUNDROOTOKODFINENESS(stepend.X);
+         stepend.Y = ROUNDROOTOKODFINENESS(stepend.Y);
          if (BSPCanMoveInRoom(Room, S, &stepend))
          {
             *P = stepend;
